+++ models/UserSearch.php
@@ -187,8 +187,6 @@ class UserSearch extends BaseUserSearch
 		]);
 
 		if (!$this->validate()) {
-			// uncomment the following line if you do not want to return any records when validation fails
-			// $query->where('0=1');
 			return $dataProvider;
 		}